.history/backend/requirements_20250908130530.txt
# 核心框架
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic_core==2.14.1
starlette==0.27.0

# 数据库
SQLAlchemy==2.0.23
mysqlclient==2.2.0
mysql-connector-python==9.4.0
redis==5.0.1
alembic==1.13.0

# AI和机器学习
openai==1.3.7
transformers==4.36.2
torch==2.1.1
numpy==1.24.3
pandas==2.1.4
scikit-learn==1.3.2
huggingface-hub==0.34.4
tokenizers==0.15.2
safetensors==0.6.2

# 自然语言处理
jieba==0.42.1
regex==2025.7.34
nltk==3.9.1
textblob==0.17.1
vaderSentiment==3.3.2

# 语音处理
SpeechRecognition==3.10.0
PyAudio @ file:///D:/project/emotion_management/backend/whl/PyAudio-0.2.11-cp311-cp311-win_amd64.whl
librosa==0.10.1
soundfile==0.13.1
audioread==3.0.1
soxr==0.5.0.post1

# 图像处理和可视化
Pillow==10.1.0
matplotlib==3.8.2
wordcloud==1.9.2
seaborn==0.13.0
contourpy==1.3.2
fonttools==4.59.2
kiwisolver==1.4.9
cycler==0.12.1

# 安全和认证
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
cryptography==41.0.7
bcrypt==4.3.0
jwt==1.4.0

# 工具库
python-dotenv==1.0.0
requests==2.31.0
httpx==0.25.2
loguru==0.7.2
aiohttp==3.9.1
websockets==10.4
psutil==5.9.6
email-validator==2.1.0
click==8.2.1
colorama==0.4.6
tqdm==4.67.1
packaging==25.0
platformdirs==4.3.8

# 科学计算
scipy==1.15.3
sympy==1.14.0
mpmath==1.3.0
numba==0.61.2
llvmlite==0.44.0
joblib==1.5.1
threadpoolctl==3.6.0

# 网络和通信
urllib3==2.5.0
certifi==2025.8.3
charset-normalizer==3.4.3
idna==3.10
dnspython==2.7.0
yarl==1.20.1
multidict==6.6.4
frozenlist==1.7.0
aiosignal==1.4.0
attrs==25.3.0
anyio==3.7.1
sniffio==1.3.1
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4

# 数据处理
python-dateutil==2.9.0.post0
pytz==2025.2
tzdata==2025.2
six==1.17.0
PyYAML==6.0.2

# 模板和配置
Jinja2==3.1.6
MarkupSafe==3.0.2
Mako==1.3.10
iniconfig==2.1.0
pluggy==1.6.0

# 文件处理
fsspec==2025.7.0
pooch==1.8.2
lazy_loader==0.4
filelock==3.19.1
watchfiles==1.1.0

# 其他依赖
annotated-types==0.7.0
typing_extensions==4.14.1
greenlet==3.2.4
decorator==5.2.1
networkx==3.5
pycparser==2.22
cffi==1.17.1
distro==1.9.0
ecdsa==0.19.1
pyasn1==0.6.1
rsa==4.9.1
boto3==1.40.24
botocore==1.40.24
jmespath==1.0.1
s3transfer==0.13.1
msgpack==1.1.1
propcache==0.3.2
win32_setctime==1.2.0

# 测试
pytest==7.4.3
pytest-asyncio==0.21.1
